% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/compute_missings.R
\name{count_missings}
\alias{count_missings}
\title{Count Missings}
\usage{
count_missings(data)
}
\arguments{
\item{data}{Dataframe in which missings will be counted}
}
\description{
Function which counts Missings in Dataframe and stores them with the Column Names
in Dataframe
}
\examples{
df <- as.data.frame(cbind(
   name = c("paul", "daniel", "paula", "beate"),
   sex = c("m", "m", "w", NA),
   income = c(1000, NA, 500, NA),
   weight = c(NA, NA, 70, NA),
   vote = c(NA, NA, NA, NA)
   ))

count_missings(df)
}
\keyword{missings}

