% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_missings.R
\name{plot_missings}
\alias{plot_missings}
\title{plot_missings}
\usage{
plot_missings(data, mis, nm, lbs, grid = "y")
}
\arguments{
\item{data}{Dataframe in which missings are stored, columns have to be named
"mi" and "p"}

\item{lbs}{Labelsize of variable names, shoul be smaller, if more variables are plotted}

\item{grid}{whether gridlines should be plotted, requieres package("cowplot")}
}
\description{
Function for plotting number of missing data,
   can be used in Conjunction with count_missings
}
\examples{
df <- as.data.frame(cbind(
   name = c("paul", "daniel", "paula", "beate"),
   sex = c("m", "m", "w", NA),
   income = c(1000, NA, 500, NA),
   weight = c(NA, NA, 70, NA),
   vote = c(NA, NA, NA, NA)
   ))

num.mis <- count_missings(df)
plot_missings(num.mis, mis = "mi", nm = "p", lbs = 10)
}
\keyword{missings,}
\keyword{plot}

